✨ Add TypeScript to the application
