+++ databases/mongodb/connect.ts
@@ -1,0 +1,11 @@
+import mongoose from 'mongoose';
+import './Models/Cinema';
+
+const connect = () => {
+  mongoose.connect('mongodb://mongodb:27017/cinema-service');
+  mongoose.connection.on('error', (err) => {
+    console.error(`Error: ${err.message}`); // eslint-disable-line
+  });
+};
+
+export default connect;
