+++ jest.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  roots: [
+    '<rootDir>/'
+  ],
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest'
+  },
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
+  moduleFileExtensions: [
+    'ts',
+    'tsx',
+    'js',
+    'json',
+    'node'
+  ]
+};
