+++ nodemon.json
@@ -1,0 +1,1 @@
+{ "watch": ["src/**/*.ts"], "ignore": ["src/**/*.spec.ts"], "exec": "ts-babel-node ./index.ts" }
