+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "./node_modules/.bin/nodemon app.js",
+    "start": "ts-node app.ts",
     "test": "./node_modules/.bin/jest --watch",
     "test-ci": "./node_modules/.bin/jest"
   },
@@ -14,8 +14,11 @@
     "collectCoverage": true
   },
   "dependencies": {
+    "@types/mongoose": "^5.0.10",
+    "babel-preset-env": "^1.6.1",
     "body-parser": "^1.18.2",
     "dotenv": "^5.0.1",
+    "es6-shim": "^0.35.3",
     "express": "^4.16.2",
     "helmet": "^3.12.0",
     "mockingoose": "^2.9.1",
@@ -25,15 +28,34 @@
     "sequelize": "^4.37.6",
     "sequelize-cli": "^4.0.0",
     "swagger-ui-express": "^3.0.0",
+    "ts-node": "^6.0.0",
+    "typescript": "^2.8.1",
     "yamljs": "^0.3.0"
   },
   "devDependencies": {
+    "@types/body-parser": "^1.16.8",
+    "@types/express": "^4.11.1",
+    "@types/helmet": "^0.0.37",
+    "@types/jest": "^22.2.3",
+    "@types/yamljs": "^0.2.30",
     "babel-eslint": "^8.2.2",
+    "babel-register": "^6.26.0",
     "eslint": "^4.18.2",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "jest": "^22.4.2",
-    "nodemon": "^1.17.2"
+    "nodemon": "^1.17.2",
+    "ts-jest": "^22.4.4"
+  },
+  "babel": {
+    "presets": [
+      "env",
+      {
+        "targets": {
+          "node": "current"
+        }
+      }
+    ]
   }
 }
