+++ tslint.json
@@ -1,0 +1,13 @@
+{
+  "defaultSeverity": "error",
+  "extends": ["tslint:recommended"],
+  "jsRules": {},
+  "rules": {
+    "quotemark": [true, "single", "jsx-double"],
+    "object-literal-sort-keys": false,
+    "arrow-parens": false,
+    "interface-name": false,
+    "object-literal-key-quotes": false
+  },
+  "rulesDirectory": []
+}
